Enabling clonediskfiles
clonediskfiles enables saving ANTsR sessions including the files that are on disk, but to which vectors of filenames point to (i.e., antsRegistration()$fwdtransforms). This allows a complete re-instantiation of the workspace on another computer, and the potential for saved sessions to be shared between collaborators without worrying about copying the files separately. It also allows finally saving and reloading antsRegistration() outputs in full.