WIP: add bkandel patch analysis
this commit is based on copying C++ only work to current ANTsR.
the source of the code is https://github.com/bkandel/PatchAnalysis
commit d90e209f4a441b635e9abc76b57b732bd0d7b471.  the plan is to
implement an R interface that generates the rotation invariant
representation and offloads the svd calculations to R.  further
clarity will be gained as work proceeds.  ben kandel and b avants
implemented the original version with ben reaching the first
well-implemented and validated version of the theory. the paper
associated with this work is PMID: 25449745.