ENH: ants joint label fusion c++ implementation
this is a much faster version of the code than the R version. there are some small issues.
the primary one is that the image types (unsigned int) create some compatibility / casting
problems that i have not worked out. for this reason, i must compute the probability maps
for the labels and recompute the final label set in order to correctly derive the output
segmentation map.  this is a BUG that should be addressed in the future.  the second
issue is that this function creates a side effect of modifying the input mask such that
only non-zero label regions are passed to JLF at the c++ level.