Fixing plot.antsImage to output a title
`title()` was called too early causing `Error in title(title.img) : plot.new has not been called yet`.
The vertical placement of the title now can be adjusted.

Note, the title works only for plots displayed in the R window. For output files, the title falls somewhere outside the plot (not displayed), and the width and height are already set according to image dimensions. So, no title visible for output images. This is a bug that is still there for now.